# Conflicts:
#	domain/src/main/java/com/baghdad/domain/repository/MovieRepository.kt
#	repository/src/main/java/com/baghdad/repository/MovieRepositoryImpl.kt
#	repository/src/main/java/com/baghdad/repository/mapper/MovieMapper.kt
#	viewmodel/src/main/java/com/baghdad/viewmodel/home/HomeViewModel.kt

+++ entity/src/main/java/com/baghdad/entity/media/TvShow.kt
@@ -7,7 +7,7 @@ data class TvShow(
     val title: String,
     val genres: List<Genre>,
     val averageRating: Double,
-    val userRating: Double?,
+    val userRating: Int?,
     val releaseDate: LocalDate,
     val overview: String,
     val posterImageURL: String,
